pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-punisher"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = ["python-chess>=1.999,<2.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
